projects/01_fyyur/starter_code/requirements.txt
alembic==1.8.1
Babel==2.10.3
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
